Fix deployment catalog CI checks
