+++ cmd/herm/models_swe.go
@@ -1,0 +1,104 @@
+// models_swe.go fetches and matches SWE-bench scores used to annotate model listings.
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"strings"
+	"time"
+)
+
+// SWE-bench leaderboard types
+
+const sweBenchURL = "https://raw.githubusercontent.com/SWE-bench/swe-bench.github.io/master/data/leaderboards.json"
+
+type sweBenchResponse struct {
+	Leaderboards []sweBenchLeaderboard `json:"leaderboards"`
+}
+
+type sweBenchLeaderboard struct {
+	Name    string           `json:"name"`
+	Results []sweBenchResult `json:"results"`
+}
+
+type sweBenchResult struct {
+	Name     string   `json:"name"`
+	Resolved float64  `json:"resolved"`
+	Tags     []string `json:"tags"`
+}
+
+// parseSWEScores extracts the highest SWE-bench Verified score per model tag
+// from leaderboard results. Returns a map from model tag identifier (e.g.
+// "claude-opus-4-5-20251101") to the best resolved score.
+func parseSWEScores(resp sweBenchResponse) map[string]float64 {
+	scores := make(map[string]float64)
+	for _, lb := range resp.Leaderboards {
+		if lb.Name != "Verified" {
+			continue
+		}
+		for _, r := range lb.Results {
+			var modelTags []string
+			for _, tag := range r.Tags {
+				if strings.HasPrefix(tag, "Model: ") {
+					modelTags = append(modelTags, strings.TrimPrefix(tag, "Model: "))
+				}
+			}
+			// Skip entries with multiple model tags (multi-model systems)
+			if len(modelTags) != 1 {
+				continue
+			}
+			tag := modelTags[0]
+			if r.Resolved > scores[tag] {
+				scores[tag] = r.Resolved
+			}
+		}
+		break // only process "Verified"
+	}
+	return scores
+}
+
+// matchSWEScoresOptions is the parameter bundle for matchSWEScores.
+type matchSWEScoresOptions struct {
+	models []ModelDef
+	scores map[string]float64
+}
+
+// matchSWEScores enriches models with SWE-bench scores by fuzzy-matching
+// model IDs against SWE-bench model tags.
+func matchSWEScores(opts matchSWEScoresOptions) {
+	models, scores := opts.models, opts.scores
+	for i := range models {
+		id := models[i].ID
+		// Try exact match first, then check if either contains the other
+		for tag, score := range scores {
+			if tag == id || strings.Contains(tag, id) || strings.Contains(id, tag) {
+				if score > models[i].SWEScore {
+					models[i].SWEScore = score
+				}
+			}
+		}
+	}
+}
+
+// fetchSWEScores fetches the SWE-bench Verified leaderboard and returns
+// a map of model tag identifiers to their best scores.
+func fetchSWEScores() (map[string]float64, error) {
+	client := &http.Client{Timeout: 15 * time.Second}
+	resp, err := client.Get(sweBenchURL)
+	if err != nil {
+		return nil, fmt.Errorf("fetching SWE-bench scores: %w", err)
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != http.StatusOK {
+		return nil, fmt.Errorf("SWE-bench API returned status %d", resp.StatusCode)
+	}
+
+	var body sweBenchResponse
+	if err := json.NewDecoder(resp.Body).Decode(&body); err != nil {
+		return nil, fmt.Errorf("decoding SWE-bench response: %w", err)
+	}
+
+	return parseSWEScores(body), nil
+}
